Make AboutFunctions.js pass
The Koan passes, but I had to use strange string formatting on Line 109 in order to have the application consider it correct. Essentially, the application is testing to see if the user knows that functions can be converted into strings. However, when filling in the answer, the application expects the string to match the whitespace exactly. If it does not, the interpreter throws an error and the entire AboutFunctions.js section gets ignored.