Implement and test Advanced Problem "find the largest prime factor of a composite number"
Breaking down my approach, my plan was to 1) use _.range and _.filter to generate an array of all unique factors, 2) reverse the array so factors go from greatest to least, and 3) iterate through the array to identify the highest prime number. For Step 3, I initially wanted to use _.reduce, but could not find a way to make it work without an excessive number of additional arrays. Instead, I tried using some very basic recursion and just plugged the highest factor back into the function. From there, if the integer could not be broken down any further, the function returns the parameter, which we expect to be the highest prime factor