Implement and test Advanced Problem "palindrome made from the product of two 3 digit numbers"
Breaking down my approach:  1) create the product and convert to a string, 2) use the second loop to go through the string and lop off the final digit each time, 3) check if the remaining string is a palindrome that is longer than the one previously detected... if so, store it in a 'result' variable, instantiated in the second line of the function, 4) back in the first loop, reset the original string and remove the first digit, 5) use that again in the second loop, and 6) return the longest palindrome. Very challenging!